Mover archivos al lugar correcto y reparar librerias de toby
